adicionado corretamente os campos globais de memory, cria estruturas para separar do term_header, modificado as threads do main e criado funcao de parsing por nome, adicionado o calculo do loadavg utilizando o campo do sysinfo

+++ include/proc_list.h
@@ -13,7 +13,7 @@ typedef struct proc_list {
 ProcList* create_proclist();
 ProcList* add(ProcList* tl, w_proc* n_proc);
 ProcList* get_lasttl(ProcList* tl);
-int get_tprocs(ProcList* tl);
+long get_tprocs(ProcList* tl);
 
 void free_proclist(ProcList* tl);
 void print_proclist(ProcList *tl, int starts_at, int max_rows);
